% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_pheatmap_legend_title}
\alias{add_pheatmap_legend_title}
\title{Adds a legend title to pheatmap legend}
\usage{
add_pheatmap_legend_title(p, legend.text = "log(n)")
}
\arguments{
\item{p}{ggplot object produced by pheatmap.}

\item{legend.grob.index}{The grob which contains the legend grob.}
}
\description{
Adds a legend title to pheatmap legend
}
\details{
Based on Mike H.'s answer from https://stackoverflow.com/questions/36852101/r-legend-title-or-units-when-using-pheatmap
}
