% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balZav_funcs.R
\name{balZav}
\alias{balZav}
\title{Getting symmetric pivot coordinates from a strictly positive matrix
Used to calculate cor_coda with rand.add = TRUE}
\usage{
balZav(x)
}
\description{
Getting symmetric pivot coordinates from a strictly positive matrix
Used to calculate cor_coda with rand.add = TRUE
}
