% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balZav_funcs.R
\name{balZavRow}
\alias{balZavRow}
\title{Getting symmetric pivot coordinates from a matrix containing 0-values
symmetric coordinates are calculated for one sample at a time using only
strictly positive signatures.
Used to calculate cor_coda with rand.add = FALSE}
\usage{
balZavRow(x)
}
\description{
Getting symmetric pivot coordinates from a matrix containing 0-values
symmetric coordinates are calculated for one sample at a time using only
strictly positive signatures.
Used to calculate cor_coda with rand.add = FALSE
}
