% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmi.R
\name{bcmi}
\alias{bcmi}
\title{Calculating bias-corrected mutual information for signature pairs}
\usage{
bcmi(x, p.val = 0.05, p.adjust = TRUE)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{p.val}{All the correlations with >=p.val are set to 0.}

\item{p.adjust}{If TRUE, p.value adjustment with BH correction is applied
before setting correlations with p.value >= p.val to 0.}
}
\value{
BCMI matrix
}
\description{
Calculating bias-corrected mutual information for signature pairs
}
