% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_metric}
\alias{calc_metric}
\title{Calculates the metric function for a matrix, by performing bootstrapping.}
\usage{
calc_metric(dt.sigs, action_func, sample.rate, sample.N, N, seed = NULL)
}
\arguments{
\item{dt.sigs}{A dataframe or a matrix of mutational signatures.}

\item{action_func}{Function calculating the interaction from a matrix}

\item{sample.rate}{The proportion of samples which should be sampled during
the bootstrapping procedure. Default: 0.9}

\item{sample.N}{Number of times to sample. Default: 100}

\item{N}{Number of times the function should be called for each sampled set:
Default: 1}

\item{seed}{Random number generator seed. Default: NULL.}
}
\description{
Calculates the metric function for a matrix, by performing bootstrapping.
}
