% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bootstrap_metric.R
\name{calculate_bootstrap_metric}
\alias{calculate_bootstrap_metric}
\title{This function calculates interaction values using a given metric_func in a
dataset through bootstrapping.}
\usage{
calculate_bootstrap_metric(
  data.input,
  metric_func,
  min.tissue.samples = 20,
  min.sig.mutcounts = 100,
  sample.rate = 0.9,
  sample.N = 100,
  N = 1,
  seed = 1,
  ...
)
}
\arguments{
\item{data.input}{Input mutational signatures}

\item{metric_func}{Function calculating the interaction from a matrix}

\item{min.tissue.samples}{Tissues with less than
this number of samples are discarded. Default: 20}

\item{min.sig.mutcounts}{Signatures with total counts less than this value
will be filtered out. Default: 100}

\item{sample.rate}{The proportion of samples which should be sampled during
the bootstrapping procedure. Default: 0.9}

\item{sample.N}{Number of times to sample. Default: 100}

\item{N}{Number of times the function should be called for each sampled set:
Default: 1}

\item{seed}{Random number generator seed. Default: 1.}

\item{...}{Parameters to be passed to metric_func.}
}
\description{
This function calculates interaction values using a given metric_func in a
dataset through bootstrapping.
}
