% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{concat_networks}
\alias{concat_networks}
\title{Summarize different metric outputs for an interaction}
\usage{
concat_networks(network.lists, tissue, filter.list, filter.mat = NULL)
}
\arguments{
\item{network.lists}{The input list.}

\item{tissue}{Tissue to be extracted.}

\item{filter.list}{A list of metric values to be filtered out. The list
element names have the same names as metric names in network.lists.
The values specify filtering threshold. Everything below the number is
set to 0. Default: NULL}

\item{filter.mat}{Passed to get_tissue_dataset_networks.}
}
\description{
Summarize different metric outputs for an interaction
}
