% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence.R
\name{cooccurrence}
\alias{cooccurrence}
\title{Pairwise signature cooccurrences across samples}
\usage{
cooccurrence(x, p.val = 0.05, maxval = 10, p.adjust = TRUE, ...)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{p.val}{All the correlations with >=p.val are set to 0.}

\item{maxval}{The value to which Inf values will be converted to.}

\item{p.adjust}{If TRUE, p.value adjustment with BH correction is applied
before setting correlations with p.value >= p.val to 0. default: TRUE}

\item{...}{arguments passed to cor.test}
}
\value{
Signature cooccurrence matrix. Values correspond to Fisher's test
odds ratios.
}
\description{
Pairwise signature cooccurrences across samples
}
\details{
For each pair of signatures cooccurrence is calculated using
Fisher's exact test. Before the test is performed, 1 is added to the table.
}
