% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_coda.R
\name{cor_coda}
\alias{cor_coda}
\title{Pairwise signature correlations from symmetric pivot coordinates}
\usage{
cor_coda(x, p.val = 0.05, rand.add = FALSE, p.adjust = TRUE, mi = FALSE, ...)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{p.val}{All the correlations with >=p.val are set to 0.}

\item{rand.add}{If rand.add = TRUE, then the matrix is strictly positive,
otherwise rand.add = FALSE(default). Check what you're}

\item{p.adjust}{If TRUE, p.value adjustment with BH correction is applied
before setting correlations with p.value >= p.val to 0.}

\item{...}{arguments passed to cor.test}
}
\value{
Signature correlation matrix
}
\description{
Pairwise signature correlations from symmetric pivot coordinates
}
\details{
For each pair of signatures correlation based on symmetric pivot
coordinates. The non-CoDa version of the function is \code{\link{cor_sigs}}.
}
