% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sigs.R
\name{cor_sigs}
\alias{cor_sigs}
\title{Simple correlation of signature}
\usage{
cor_sigs(x, p.val = 0.05, p.adjust = TRUE, ...)
}
\arguments{
\item{x}{Data.frame or a matrix of signature intensities(or counts).
Signatures need to correspond to columns.}

\item{p.val}{All the correlations with >=p.val are set to 0.}

\item{p.adjust}{If TRUE, p.value adjustment with BH correction is applied
before setting correlations with p.value >= p.val to 0.}

\item{...}{arguments passed to cor.test}
}
\value{
Signature correlation matrix
}
\description{
Simple correlation of signature
}
\details{
All pairwise correlations are calculated by cor.test.
Non-CoDa version of \code{\link{cor_coda}} function, provided for comparison.
}
