% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_common_sigs}
\alias{get_common_sigs}
\title{The function summarizes interactions for a given metric across tissues,
identifies common interaction motifs.}
\usage{
get_common_sigs(metric.list, metric, outdir, threshold = 0.2)
}
\arguments{
\item{metric.list}{input metric.list, element names are different interaction
metrics. Each element is a list corresponding to interactions corresponding
to individual tissues.}

\item{metric}{Metric for which the summaries should be obtained.}

\item{outdir}{The output directory where the interaction matrices and the
outputs should be written.}

\item{threshold}{All interactions below this threshold are set to 0.
Default: 0.2.}
}
\description{
The function summarizes interactions for a given metric across tissues,
identifies common interaction motifs.
}
