% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_interaction_tissues}
\alias{get_interaction_tissues}
\title{For a list of interaction metrics this function summarizes the positive and
negative interactions in a matrix, where for each interaction the cell
contains a comma-separated tissue  names where this interaction was observed.}
\usage{
get_interaction_tissues(list.of.int.elems, threshold = 0.1)
}
\arguments{
\item{list.of.int.elems}{A list with matrix elements.}

\item{threshold}{All the values below the threshold are discarded.}
}
\value{
ggplot object
}
\description{
For a list of interaction metrics this function summarizes the positive and
negative interactions in a matrix, where for each interaction the cell
contains a comma-separated tissue  names where this interaction was observed.
}
\details{
The input is a list of matrices corresponding to e.g. interactions
in different tissues. The output plot summarizes the number of list elements
in which this interaction happens either in positive or negative direction.
The positive and negative interactions are summarized separately.
}
