% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_relevant_clin_df}
\alias{get_relevant_clin_df}
\title{Get survival data for relevant tissues}
\usage{
get_relevant_clin_df(clin.df, dataset, is.TCGA, tissues)
}
\arguments{
\item{clin.df}{Clinical data for samples. E.g. PCAWG.clin.df.}

\item{dataset}{Signature-activities for samples. E.g.PCAWG.full.subset.ann.
The first three columns are Cancer.Types, Sample.Names, Accuracy and fourth on
signature activities.}

\item{is.TCGA}{Indicate if TCGA data is used}

\item{tissues}{A vector of tissues where the interaction of signatures should
be assessed.}
}
\description{
Get survival data for relevant tissues
}
