% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sample_sheet}
\alias{get_sample_sheet}
\title{A function-wrapper to read sample sheet.}
\usage{
get_sample_sheet(file.dir)
}
\arguments{
\item{file.dir}{File directory.}
}
\description{
A function-wrapper to read sample sheet.
}
\details{
The function can be later used in case some sort of processing
becomes necessary.
}
