% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sig_path_assocs}
\alias{get_sig_path_assocs}
\title{For a given signature activity matrix and pathway alteration matrix
computes Fisher's exact test for all pathway-signature interactions.}
\usage{
get_sig_path_assocs(
  sigs.df,
  pathways.df,
  p.val.threshold = 0.05,
  p.adjust = TRUE,
  method = "BH"
)
}
\arguments{
\item{sigs.df}{signatures dataframe or matrix}

\item{pathways.df}{pathways dataframe or matrix}

\item{p.val.threshold}{p-value threshold for BH correction. Default: 0.05}

\item{p.adjust}{Controls if p-values should be adjusted. Default: TRUE}

\item{method}{P-value adjustement methods. Default: BH}
}
\value{
A matrix with signature-number of rows and pathway-number of columns
with values indicating odds ratio of corresponding signatures and pathways.
}
\description{
For a given signature activity matrix and pathway alteration matrix
computes Fisher's exact test for all pathway-signature interactions.
}
