% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sig_path_lms}
\alias{get_sig_path_lms}
\title{For signature and pathway matrices performs linear regression for
each signature-pathway pair.
For a given signature activity matrix and pathway alteration matrix
computes Fisher's exact test for all pathway-signature interactions.}
\usage{
get_sig_path_lms(
  sigs.df,
  pathways.df,
  sig.log = TRUE,
  robust = TRUE,
  path.to.sig = TRUE,
  p.val.threshold = 0.05,
  p.adjust = TRUE,
  method = "BH",
  ...
)
}
\arguments{
\item{sigs.df}{signatures dataframe or matrix}

\item{pathways.df}{pathways dataframe or matrix}

\item{sig.log}{Defines whether signatures should be logged. Default: TRUE}

\item{robust}{Specifies if robust statistics should be used.
Robust functions are called from robustbase package. Default: TRUE}

\item{path.to.sig}{If true, linear regression is used to predict signatures
from pathways. Otherwise, logistic regression is used to predict pathways
from signatures. Default: TRUE}

\item{p.val.threshold}{p-value threshold for BH correction. Default: 0.05}

\item{p.adjust}{Controls if p-values should be adjusted. Default: TRUE}

\item{method}{P-value adjustement methods. Default: BH}
}
\value{
A matrix with signature-number of rows and pathway-number of columns
with values indicating odds ratio of corresponding signatures and pathways.
}
\description{
For signature and pathway matrices performs linear regression for
each signature-pathway pair.
For a given signature activity matrix and pathway alteration matrix
computes Fisher's exact test for all pathway-signature interactions.
}
