% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sig_stars}
\alias{get_sig_stars}
\title{Significance starts from a pvalue.}
\usage{
get_sig_stars(p.val)
}
\arguments{
\item{p.val}{Input p value}
}
\description{
Significance starts from a pvalue.
}
