% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_specimen_hist}
\alias{get_specimen_hist}
\title{A function-wrapper to read specimen histology file.}
\usage{
get_specimen_hist(file.dir)
}
\arguments{
\item{file.dir}{File directory.}
}
\description{
A function-wrapper to read specimen histology file.
}
\details{
The function can be later used in case some sort of processing
becomes necessary.
}
