% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surv_best_model}
\alias{get_surv_best_model}
\title{Running batch survival analysis tests for a set of interactions. Returns a
list of cox regression outputs with the best selected model. If the tissue
supports a survival effect for a given interaction, then this model is selected.
Among several models, the one with the lowest loglik is .}
\usage{
get_surv_best_model(
  sig.sig.tissues.matrix,
  dataset,
  clin.df,
  param.list,
  min.sample.fraction = 0
)
}
\arguments{
\item{sig.sig.tissues.matrix}{A matrix with rows and columns with signatures,
and the elements are comma-separated tissue names where that interaction is
observed.
@param dataset The signature values for all samples. E.g. PCAWG.full.subset.ann
@param clin.df The dataframe with clinical info.
@param with.total.muts If TRUE the total number of mutations in the samples will
be provided as a confounder to the model. Default: TRUE
@param tmb.logged If TRUE the tumor mutational burden will be logged.
Default: TRUE
@param binary.status If TRUE, the model will compare samples with both signatures
with all the other samples having either of the signatures or none. Default:
FALSE
@export}
}
\description{
Running batch survival analysis tests for a set of interactions. Returns a
list of cox regression outputs with the best selected model. If the tissue
supports a survival effect for a given interaction, then this model is selected.
Among several models, the one with the lowest loglik is .
}
