% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surv_bestcoxlist}
\alias{get_surv_bestcoxlist}
\title{Running batch survival analysis tests for a set of interactions. Returns a
list of best cox regression outputs for all the interaction tests
which didn't fail. The models are selected based on whether they show an
effect for the interaction and among those models with one with the highest
loglik is picked.}
\usage{
get_surv_bestcoxlist(
  sig.sig.tissues.matrix,
  dataset,
  clin.df,
  with.total.muts = TRUE,
  tmb.logged = TRUE,
  binary.status = FALSE
)
}
\arguments{
\item{sig.sig.tissues.matrix}{A matrix with rows and columns with signatures,
and the elements are comma-separated tissue names where that interaction is
observed.
@param dataset The signature values for all samples. E.g. PCAWG.full.subset.ann
@param clin.df The dataframe with clinical info.
@param with.total.muts If TRUE the total number of mutations in the samples will
be provided as a confounder to the model. Default: TRUE
@param tmb.logged If TRUE the tumor mutational burden will be logged.
Default: TRUE
@param binary.status If TRUE, the model will compare samples with both signatures
with all the other samples having either of the signatures or none. Default:
FALSE}
}
\description{
Running batch survival analysis tests for a set of interactions. Returns a
list of best cox regression outputs for all the interaction tests
which didn't fail. The models are selected based on whether they show an
effect for the interaction and among those models with one with the highest
loglik is picked.
}
