% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_surv_plotlist}
\alias{get_surv_plotlist}
\title{Running batch survival analysis tests for a set of interactions. Returns a
plotlist for all the interaction test which didn't fail.}
\usage{
get_surv_plotlist(
  sig.sig.tissues.matrix,
  dataset,
  clin.df,
  age.at.diagnosis = TRUE,
  with.total.muts = TRUE,
  tmb.logged = TRUE,
  binary.status = FALSE,
  epistatic = FALSE,
  legend.pos = c(0.8, 0.8)
)
}
\arguments{
\item{sig.sig.tissues.matrix}{A matrix with rows and columns with signatures,
and the elements are comma-separated tissue names where that interaction is
observed.}

\item{dataset}{The signature values for all samples. E.g. PCAWG.full.subset.ann}

\item{clin.df}{The dataframe with clinical info.}

\item{with.total.muts}{If TRUE the total number of mutations in the samples will
be provided as a covariate to the model. Default: TRUE}

\item{tmb.logged}{If TRUE the tumor mutational burden will be logged.
Default: TRUE}

\item{binary.status}{If TRUE, the model will compare samples with both signatures
with all the other samples having either of the signatures or none. Default:
FALSE}

\item{epistatic}{The survival model tests the interaction of int1 and int2 as
int1 * int2. Default: FALSE}

\item{legend.pos}{Position of the legend in the survival plot.}
}
\description{
Running batch survival analysis tests for a set of interactions. Returns a
plotlist for all the interaction test which didn't fail.
}
