% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_dataset_networks.R
\name{get_tissue_dataset_networks}
\alias{get_tissue_dataset_networks}
\title{Extracts the interaction metrics of a tissue
from a list of interaction networks.}
\usage{
get_tissue_dataset_networks(
  tissue,
  network.lists,
  filter.list = NULL,
  filter.mat = NULL
)
}
\arguments{
\item{tissue}{Tissue to be extracted}

\item{network.lists}{The input list}

\item{filter.list}{A list of metric values to be filtered out. The list
element names have the same names as metric names in network.lists.
The values specify filtering threshold. Everything below the number is
set to 0. Default: NULL}

\item{filter.mat}{If provided, this is a list of matrices, where the name of
the element is the metric name. The list elements are matrices of all
signature combinations in that tissue, where matrix elements are a predefined
quantile of the null model of given sig-sig interaction. If provided the
matrix will add another layer of filtering, where the absolute metric values
smaller than the provided value are set to 0.}
}
\value{
A list with interaction metric matrices. The list element names match
those in the input network.lists.
}
\description{
Extracts the interaction metrics of a tissue
from a list of interaction networks.
}
