% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_tissue_pathway_activities}
\alias{get_tissue_pathway_activities}
\title{Extract signature and pathway activities for a tissue}
\usage{
get_tissue_pathway_activities(tissue, sigs.input, pathways.input)
}
\arguments{
\item{tissue}{Tissue to extract}

\item{sigs.input}{Signature activities}

\item{pathways.input}{Pathway activities}
}
\value{
A list with sigs and pathways elements with corresponding matrices.
Can be forwarded do interaction calculting functions
}
\description{
Extract signature and pathway activities for a tissue
}
