% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{A function to generate default colors in ggplots.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors.}
}
\description{
A function to generate default colors in ggplots.
}
