% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggheatmap_wrapper}
\alias{ggheatmap_wrapper}
\title{A ggheatmap wrapper for interaction matrix returns}
\usage{
ggheatmap_wrapper(
  metric.matrix,
  colorPalette = "RdBu",
  points = T,
  revColors = F,
  revRow = T,
  scaleName = "log(OR)",
  title = NULL,
  limits = NULL,
  ...
)
}
\arguments{
\item{metric.matrix}{the input matrix}
}
\value{
Returns a ggheatmap (which is a gg object,
but not a conventional gg object unfortunately.)
}
\description{
All other parameters are passed to ggheatmap
\link[ggheatmap]{ggheatmap}
}
