% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graph_unique_edges}
\alias{graph_unique_edges}
\title{Removes identical edges between two nodes}
\usage{
graph_unique_edges(input.graph, summary.col = NULL)
}
\arguments{
\item{input.graph}{Input graph}

\item{summary.col}{This variable indicates the column, which should be summarized.
When specified, all the edges between the same nodes, which have different values
for this column are summarized, a new edge variable \code{count} is added and
the values of the specificed column are collapsed and stored in the column called
\code{summedList}}
}
\description{
Removes identical edges between two nodes
}
