% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myggheatmap.R
\name{myggheatmap}
\alias{myggheatmap}
\title{Generate a ggplot2 heatmap with row and column dendrograms
modified from https://github.com/tleonardi/ggheatmap}
\usage{
myggheatmap(
  dataMatrix,
  orderCol = T,
  orderRow = T,
  points = F,
  dendroLineSize = 0.5,
  revCol = F,
  revRow = F,
  rowLabels = NA,
  colLabels = NA,
  fontSize = 20,
  vlines = FALSE,
  vlinecol = NULL,
  vlinesize = NULL,
  viridis = FALSE,
  colorPalette = "default",
  scaleName = "value",
  distMethod = "euclidean",
  clustMethod = "complete",
  revColors = F,
  title = NULL,
  ...
)
}
\arguments{
\item{dataMatrix}{A data.frame containing the input data.}

\item{orderCol}{Reorder the columns (default=T)}

\item{orderRow}{Reorder the rows (default=T)}

\item{dendroLineSize}{Size of the dendrogram lines (default=0.5)}

\item{revCol}{reverse the column order}

\item{revRow}{revers the row order}

\item{fontSize}{Font size (default=20)}

\item{vlines}{Add vertical lines (default = FALSE)}

\item{vlinecol}{Color of vertical lines if specified}

\item{vlinesize}{Size of vertical colors if specified}

\item{colorPalette}{Color palette (default='Spectral')}

\item{scaleName}{Name of the colorscale (default='value')}

\item{distMethod}{Distance method (default='euclidean', see ?dist)}

\item{clustMethod}{Clustering method (default='complete', see ?hclust)}

\item{revColors}{Invert color scale}

\item{title}{Title of the plot (default=NULL)}
}
\description{
Generate a ggplot2 heatmap with row and column dendrograms
modified from https://github.com/tleonardi/ggheatmap
}
\examples{
myggheatmap(mtcars)
}
