% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{order_matrix_rc}
\alias{order_matrix_rc}
\title{Order matrix/dataframe columns/rows in increasing order}
\usage{
order_matrix_rc(input.mat)
}
\arguments{
\item{input.mat}{input data}
}
\description{
Order matrix/dataframe columns/rows in increasing order
}
