% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{p_matrix_process}
\alias{p_matrix_process}
\title{Performs p-value adjustment on a matrix of p-values based on input parameters.}
\usage{
p_matrix_process(p.values, p.adjust = TRUE, method = "BH")
}
\value{
A matrix with the same dimensions and dimnames as the input
}
\description{
Performs p-value adjustment on a matrix of p-values based on input parameters.
}
