% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pathways_signatures_heatmap}
\alias{pathways_signatures_heatmap}
\title{Plots a signature heatmap with pathway annotations}
\usage{
pathways_signatures_heatmap(
  tissue,
  signatures,
  pathways,
  border_color = "grey60",
  ...
)
}
\arguments{
\item{tissue}{Tissue}

\item{signatures}{Signatures dataframe, signatures start from column 4.
Column 2 is Sample.Names.}

\item{pathways}{Pathway mutations dataframe, pathways start from column 4,
contains donor_id.}

\item{border_color}{Border color of cells. This parameter is there to control
removing border color with NA. Pheatmap doesn't properly remove it
and the border is still present when saving the plot. Default: gray60.
Should be NA to remove the border_color.}

\item{...}{Params passed to pheatmap.}
}
\description{
Plots a signature heatmap with pathway annotations
}
