% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pcawg_pathway_profiles}
\alias{pcawg_pathway_profiles}
\title{The function outputs mutational profiles for mutations in genes
in the pathway of interest and for all other mutations.}
\usage{
pcawg_pathway_profiles(
  pathway,
  tissue,
  pathways.df,
  vcf.dir = here("data/raw/PCAWG/vcfs/consensus_snv_indel/snv_mnv/")
)

pcawg_pathway_profiles(
  pathway,
  tissue,
  pathways.df,
  vcf.dir = here("data/raw/PCAWG/vcfs/consensus_snv_indel/snv_mnv/")
)
}
\arguments{
\item{pathway}{One of the 10 oncogenic pathway names.}

\item{tissue}{The tissues where the profiles have to be assessed.}

\item{pathways.df}{The data.frame with sample names and pathway alterations.}

\item{vcf.dir}{Directory to vcf files}
}
\value{
A list with mutational profiles for all mutations and for those in
the pathway of interest.

A list with mutational profiles for all mutations and for those in
the pathway of interest.
}
\description{
The function outputs mutational profiles for mutations in genes
in the pathway of interest and for all other mutations.

The function outputs mutational profiles for mutations in genes
in the pathway of interest and for all other mutations.
}
