% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{percentile}
\alias{percentile}
\title{Get the percentile of a value v in a vector x.}
\usage{
percentile(x, v)
}
\arguments{
\item{x}{The vector}

\item{v}{The value}
}
\description{
Get the percentile of a value v in a vector x.
}
