% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pick_survival_model_int}
\alias{pick_survival_model_int}
\title{Tests a range of models}
\usage{
pick_survival_model_int(
  dataset,
  tissue,
  signatures,
  clin.df,
  param.values,
  min.sample.fraction = 0,
  filename = NULL,
  rm.non.sig.sheets = TRUE,
  return.only.sig = TRUE,
  sig.threshold = 0.05
)
}
\description{
Tests a range of models
}
