% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_all_counts}
\alias{plot_all_counts}
\title{For a list of interaction metrics this function summarizes the positive and
negative interactions in a plot.}
\usage{
plot_all_counts(
  list.of.int.elems,
  threshold = 0.1,
  min.abssum = 1,
  psize = 8,
  lsize = 2,
  expand.mult = c(0.04, 0.04),
  upper.triangle.only = TRUE
)
}
\arguments{
\item{list.of.int.elems}{A list with matrix elements.}

\item{threshold}{All the values below the threshold are discarded.}

\item{min.abssum}{Rows and columns containing with less than this number of
interactions (absolute) are removed. Default: 1.}

\item{psize}{Controls the size of the triangles. Default: 8.}

\item{lsize}{Label size. Default: 2.}

\item{expand.mult}{A vector of expansion factors for x and y axes.
Default. c(0.04, 0.04)}

\item{upper.triangle.only}{If true, for symmetric matrices only diagonal will be printed.
Default:TRUE.}
}
\value{
ggplot object
}
\description{
For a list of interaction metrics this function summarizes the positive and
negative interactions in a plot.
}
\details{
The input is a list of matrices corresponding to e.g. interactions
in different tissues. The output plot summarizes the number of list elements
in which this interaction happens either in positive or negative direction.
The positive and negative interactions are summarized separately.
}
