% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_bipartite2}
\alias{plot_bipartite2}
\title{Bipartite network test}
\usage{
plot_bipartite2(tissue.odds.ratios)
}
\arguments{
\item{tissue.odds.ratios}{Dataframe returned by summarize_ints_to_df}
}
\value{
gggraph with sugiyama layout
}
\description{
Bipartite network test
}
