% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_HR_vars}
\alias{plot_HR_vars}
\title{(plot hazard ratio for variables)
Makes a forest plot for hazard ratios for given parameter}
\usage{
plot_HR_vars(
  all.conds.df,
  param,
  average = TRUE,
  log.HR = FALSE,
  no_stripes = TRUE
)
}
\arguments{
\item{all.conds.df}{Data.frame with all the parameters, hazard ratios,
confidence intervals. Generated with function HR_summary_for_all.}

\item{param}{Parameter to be plotted}

\item{average}{If TRUE the parameter values for a given tissues should be
averaged. Default:TRUE.}

\item{log.HR}{If TRUE the Hazard ratio will be logged. Default: FALSE.}
}
\description{
(plot hazard ratio for variables)
Makes a forest plot for hazard ratios for given parameter
}
