% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_mixed_layout}
\alias{plot_mixed_layout}
\title{Plot interaction network with a mixed layout, where the clock-like
signatures are in the middle and all the other signatures are on a circular
layout.}
\usage{
plot_mixed_layout(
  graph.input,
  central.nodes = c("Ageing", "SBS5", "SBS40"),
  circular.node.order = NULL,
  edge.width.breaks = NULL,
  lbm = margin(0, 0, 0, 1.1, unit = "in")
)
}
\arguments{
\item{graph.input}{The network of interactions as a tbl_graph}

\item{central.nodes}{The nodes that should be in the center.
Default: c("Ageing" , "SBS5", "SBS40")}

\item{circular.node.order}{Can be used to specify the order of the nodes in the
outer layer. Default: NULL}

\item{edge.width.breaks}{Can be used to manually specify the edge widths.
Default: NULL}

\item{lbm}{legend.box.margin to be passed to theme}
}
\value{
a ggraph plot of the interaction network.
}
\description{
Plot interaction network with a mixed layout, where the clock-like
signatures are in the middle and all the other signatures are on a circular
layout.
}
\details{
Beware!!! The node colors are predefined for different etiology groups.
The etiology groups are derived from signature.annotations object
}
