% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_multi_network}
\alias{plot_multi_network}
\title{Wrapper of tissue interaction network plot}
\usage{
plot_multi_network(network.lists, tissue, filter.list, layout = "stress")
}
\arguments{
\item{network.lists}{The input list}

\item{tissue}{Tissue to be extracted}

\item{filter.list}{A list of metric values to be filtered out. The list
element names have the same names as metric names in network.lists.
The values specify filtering threshold. Everything below the number is
set to 0. Default: NULL}

\item{layout}{layout to be passed to ggraph. Default: stress. Some useful
alternative is dh.}
}
\description{
Wrapper of tissue interaction network plot
}
