% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_param_piechart}
\alias{plot_param_piechart}
\title{Creates summaries from survival summaries generated by HR_summary_for_all}
\usage{
plot_param_piechart(data, param)
}
\arguments{
\item{data}{summary output created by HR_summary_for_all}

\item{param}{the name of the parameter to summarize}
}
\description{
Creates summaries from survival summaries generated by HR_summary_for_all
}
