% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_sigint_forest}
\alias{plot_sigint_forest}
\title{Make forest plot for all the significant interaction paramters from the
survival models.}
\usage{
plot_sigint_forest(data, log.HR = TRUE)
}
\arguments{
\item{data}{The summary data across tissues and
models generated with HR_summary_for_all.}

\item{log.HR}{Indicated whether the hazard ratio is logged or not.}
}
\description{
Make forest plot for all the significant interaction paramters from the
survival models.
}
