% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multi_graphs.R
\name{print_multi_graphs}
\alias{print_multi_graphs}
\title{Plots a network of interaction metrics}
\usage{
print_multi_graphs(tissue_graphs, layout = "kk", ...)
}
\arguments{
\item{tissue_graphs}{tbl_graph generated by tissue_multi_graph}

\item{layout}{layout to be passed to ggraph. Default: kk.
Some useful alternative is dh.}
}
\description{
Plots a network of interaction metrics
}
