% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rm_zeros}
\alias{rm_zeros}
\title{Remove the columns and rows of the dataframe with all 0's}
\usage{
rm_zeros(x)
}
\arguments{
\item{x}{input dataframe}
}
\description{
Remove the columns and rows of the dataframe with all 0's
}
