% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_GES_octave_wrapper}
\alias{run_GES_octave_wrapper}
\title{A wrapper to call GES scoring tool from within R.}
\usage{
run_GES_octave_wrapper(mat, tmpdir = here("tmp"))
}
\arguments{
\item{mat}{Input matrix or a data frame - the feature matrix}

\item{tmpdir}{Directory where the temporary files should be written.}
}
\value{
DAG adjancency matrix from GES method.
}
\description{
A wrapper to call GES scoring tool from within R.
}
\details{
Initial implementation in git repo:
Biwei-Huang/Generalized-Score-Functions-for-Causal-Discovery/
G(i,j) = 1: i->j (the edge is from i to j)
G(i,j) = -1: i-j (the direction between i and j is not determined)
G(i,j) = 0:  i j (no causal edge between i and j)
}
