% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_signature_annotation}
\alias{set_signature_annotation}
\title{The function aggregates mutations based on annotations}
\usage{
set_signature_annotation(annotations, signature.matrix.df)
}
\arguments{
\item{annotations}{annotations data frame. It contains must contain 2 columns-
Signature and Annotation}

\item{signature.matrix.df}{Signature matrix, which should be annotated}
}
\value{
A matrix with the same number of rows as the input
siganture.matrix.df and aggregated and renamed columns
according to annotations dataframe.
}
\description{
The function aggregates mutations based on annotations
}
