% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sig_pathway_int_null}
\alias{sig_pathway_int_null}
\title{Assess signature-pathway interactions null model across tissues for a custom function}
\usage{
sig_pathway_int_null(
  sigs.input,
  pathways.input,
  interaction.functions,
  path.min.tissues = 30,
  N = 1000
)
}
\arguments{
\item{sigs.input}{Signature activities. The rows correspond to samples, the
first three columns are Cancer.Types, Sample.Names, Accuracy, all the other
columns correspond to signature names. The values are signature activities.}

\item{pathways.input}{Pathway status - formatted like mutated.pathways.tissues
Pathway activities start at column 4. The first three columns are sample_id,
donor_id, Cancer.Types. Other columns correspond to pathway names. The values
correspond to number of mutations in the pathway.}

\item{path.min.tissues}{Minimal number of samples in each tissue to be considered}

\item{N}{number of times the shuffling should take place}

\item{interaction_functions}{The functions defining the metrics.
E.g. get_sig_path_assocs}
}
\value{
A list with the length of abundant tissues in the datasets,
where each element is a 3D array, where first two dims are sig and path, the 3rd
dim is the N metric values from N random shufflings.
}
\description{
Assess signature-pathway interactions null model across tissues for a custom function
}
