% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature.annotations}
\alias{signature.annotations}
\title{Signature annotations based on cosmic and curated by
hakobyan et al., 2023}
\format{
A dataframe with 72 rows and 5 columns:
\describe{
\item{Signature}
\item{Class}
\item{Description}
\item{Annotation}
\item{Origin}
}
}
\usage{
signature.annotations
}
\description{
Signature annotations based on cosmic and curated by
hakobyan et al., 2023
}
\examples{
head(signature.annotations)
}
\keyword{datasets}
