% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subset_tissue}
\alias{subset_tissue}
\title{Extract tissue and the signatures active in that tissue from a matrix of all
the data.}
\usage{
subset_tissue(sigs.full.subset, tissue)
}
\arguments{
\item{sigs.full.subset}{is the signature activity matrix, where first column
is the cancer type/tissue and signature activities start from column 4. This
is the input type for signature activities from Alexandrov et al.(Nature, 2020)
paper.}

\item{tissue}{Tissue of interest}
}
\value{
Returns a subset of samples from the selected tissue and only active
signatures
}
\description{
Extract tissue and the signatures active in that tissue from a matrix of all
the data.
}
