% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_by_tissues}
\alias{summarize_by_tissues}
\title{Summarize positive and negative interactions across tissues}
\usage{
summarize_by_tissues(summary.list, tissue.names = FALSE)
}
\arguments{
\item{summary.list}{A list of interactions for individual tissues. Each
element is a matrix.}

\item{tissue.names}{Input tissue names}
}
\value{
a list of two matrices for each dataset. The elements of the matrices
are lists as elements with two elements called pos and neg for positive and
negative interactions.
}
\description{
For each signature pair the function counts how many positive and negative
interactions were observed across tissues and returns a matrix of lists with
two elements called pos and neg or NULL if no interactions were found.
}
