% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_int_mat}
\alias{summarize_int_mat}
\title{Summarizing individual matrices of interactions}
\usage{
summarize_int_mat(interaction.list, pos.ints = TRUE)
}
\arguments{
\item{interaction.list}{Input list}

\item{pos.ints}{Defines if positive or negative interactions
should be considered. Default: TRUE}
}
\value{
Matrix with element is concatenated tissue names
bearing the interaction
}
\description{
Summarizing individual matrices of interactions
}
