% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_ints_to_df}
\alias{summarize_ints_to_df}
\title{For a list of interaction metrics this function summarizes the positive and
negative interactions into a dataframe which can then be fed into plotting functions.}
\usage{
summarize_ints_to_df(
  list.of.int.elems,
  threshold = 0.1,
  min.abssum = 1,
  upper.triangle.only = TRUE
)
}
\arguments{
\item{list.of.int.elems}{A list with matrix elements.}

\item{threshold}{All the values below the threshold are discarded.}

\item{min.abssum}{Rows and columns containing with less than this number of
interactions (absolute) are removed. Default: 1.}

\item{upper.triangle.only}{If true, for symmetric matrices only upper triangle will be printed.
Default:TRUE.}
}
\value{
Dataframe for plotting.
}
\description{
For a list of interaction metrics this function summarizes the positive and
negative interactions into a dataframe which can then be fed into plotting functions.
}
