% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_ints_to_df_new}
\alias{summarize_ints_to_df_new}
\title{For a list of interaction metrics this function summarizes the positive and
negative interactions, aggregates tissue names.}
\usage{
summarize_ints_to_df_new(int.value.list, threshold = 0.1, min.abssum = 1)
}
\arguments{
\item{int.value.list}{A list with matrix elements.}

\item{threshold}{All the values below the threshold are discarded.}

\item{min.abssum}{Rows and columns containing with less than this number of
interactions (absolute) are removed. Default: 1.}
}
\value{
Dataframe for plotting.
}
\description{
For a list of interaction metrics this function summarizes the positive and
negative interactions, aggregates tissue names.
}
