% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{surv_add_param_fillers}
\alias{surv_add_param_fillers}
\title{For survival summaries generated by HR_summary_for_all, if a parameter is not
available for a condition, the parameter is added with zero-values for overall
statistics.}
\usage{
surv_add_param_fillers(dinput, param)
}
\arguments{
\item{param}{the name of the parameter to fill saturate the data with}

\item{data}{summary output created by HR_summary_for_all}
}
\description{
For survival summaries generated by HR_summary_for_all, if a parameter is not
available for a condition, the parameter is added with zero-values for overall
statistics.
}
