% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survival_for_covariates}
\alias{survival_for_covariates}
\title{Survival analysis for covariates with Cox proportional
hazards regression model.}
\usage{
survival_for_covariates(
  dataset,
  clin.df,
  tissue,
  age.at.diagnosis = FALSE,
  with.total.muts = FALSE,
  tmb.logged = TRUE,
  is.TCGA = FALSE
)
}
\arguments{
\item{dataset}{Signature-activities for samples. E.g.PCAWG.full.subset.ann.
The first three columns are Cancer.Types, Sample.Names, Accuracy and fourth on
signature activities.}

\item{clin.df}{Clinical data for samples. E.g. PCAWG.clin.df.}

\item{tissue}{A vector of tissues where the interaction of signatures should
be assessed.}

\item{with.total.muts}{If TRUE the total number of mutations in the samples will
be provided as a covariate to the model. Default: TRUE}

\item{tmb.logged}{If TRUE the tumor mutational burden will be logged.
Default: TRUE}

\item{is.TCGA}{TRUE for TCGA}

\item{signatures}{A vector of two signatures which should be considered for
survival analysis.}
}
\description{
Survival analysis for covariates with Cox proportional
hazards regression model.
}
