% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survival_for_interactions}
\alias{survival_for_interactions}
\title{Survival analysis for signature-signature interactions with Cox proportional
hazards regression model.}
\usage{
survival_for_interactions(
  dataset,
  clin.df,
  signatures,
  tissues,
  legend_pos = c(0.8, 0.8),
  age.at.diagnosis = TRUE,
  with.total.muts = TRUE,
  tmb.logged = TRUE,
  binary.status = FALSE,
  epistatic = FALSE,
  get_df = FALSE,
  conf.int = FALSE
)
}
\arguments{
\item{dataset}{Signature-activities for samples. E.g.PCAWG.full.subset.ann.
The first three columns are Cancer.Types, Sample.Names, Accuracy and fourth on
signature activities.}

\item{clin.df}{Clinical data for samples. E.g. PCAWG.clin.df.}

\item{signatures}{A vector of two signatures which should be considered for
survival analysis.}

\item{tissues}{A vector of tissues where the interaction of signatures should
be assessed.}

\item{legend_pos}{Legend position. Default: c(0.8, 0.8).}

\item{with.total.muts}{If TRUE the total number of mutations in the samples will
be provided as a confounder to the model. Default: TRUE}

\item{tmb.logged}{If TRUE the tumor mutational burden will be logged.
Default: TRUE}

\item{binary.status}{If TRUE, the model will compare samples with both signatures
with all the other samples having either of the signatures or none. Default:
FALSE}

\item{epistatic}{The survival model tests the interaction of int1 and int2 as
int1 * int2. Default: FALSE}

\item{get_df}{If true, the function returns the constructed model dataframe
(for debugging). Default:FALSE.}

\item{conf.int}{If TRUE confidence interval should be added to ggsurvplot.}
}
\description{
Survival analysis for signature-signature interactions with Cox proportional
hazards regression model.
}
