% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_CpG}
\alias{test_CpG}
\title{Hypergeometric test for C>T mutations at CpG sites between two profiles.}
\usage{
test_CpG(seven_channel_profile1, seven_channel_profile2, lower.tail = TRUE)

test_CpG(seven_channel_profile1, seven_channel_profile2, lower.tail = TRUE)
}
\arguments{
\item{seven_channel_profile1}{First profile}

\item{seven_channel_profile2}{Second profile}

\item{lower.tail}{Parameter will be passed to phyper}
}
\value{
P value or 1-P of a hypergeometric test.

P value or 1-P of a hypergeometric test.
}
\description{
Hypergeometric test for C>T mutations at CpG sites between two profiles.

Hypergeometric test for C>T mutations at CpG sites between two profiles.
}
