% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_multi_graph.R
\name{tissue_multi_graph}
\alias{tissue_multi_graph}
\title{Creates a tbl_graph network from a list of calculated metrics.}
\usage{
tissue_multi_graph(network_list, network_names = NULL)
}
\arguments{
\item{network_list}{A list of calculated metrics matrices}

\item{network_names}{Names of metrics can be overriden with this parameter.}
}
\description{
Creates a tbl_graph network from a list of calculated metrics.
}
\examples{
mat = matrix(sample(100, 300, replace = TRUE), 30, 10,
		   dimnames = list(paste0("c", 1:30), letters[1:10]))
mat[, 2] = 2 * mat[,1] + mat[,2]
mat[runif(300) < 0.5] = 0


metric_functions = list(CoDa = cor_coda,
                        cooccurrence = cooccurrence,
                        MI = bcmi,
                        Spearman = function(x) {cor_sigs(x, method = "spearman")})

metrics_out = lapply(metric_functions, function(x) x(mat))
graph_out = tissue_multi_graph(metrics_out)

}
