% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tissue_dataset_networks.R
\name{get_tissue_dataset_networks}
\alias{get_tissue_dataset_networks}
\title{Extracts the interaction metrics of a tissue
from a list of interaction networks.}
\usage{
get_tissue_dataset_networks(tissue, network.lists, filter.list = NULL)
}
\arguments{
\item{tissue}{Tissue to be extracted}

\item{network.lists}{The input list}

\item{filter.list}{A list of metric values to be filtered out. The list
element names have the same names as metric names in network.lists.
The values specify filtering threshold. Everything below the number is
set to 0. Default: NULL}
}
\value{
A list with interaction metric matrices. The list element names match
those in the input network.lists.
}
\description{
Extracts the interaction metrics of a tissue
from a list of interaction networks.
}
