% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pick_survival_model_int}
\alias{pick_survival_model_int}
\title{Tests a range of models}
\usage{
pick_survival_model_int(
  dataset = dataset,
  signatures = c(sig1, sig2),
  tissues = tissue,
  clin.df = clin.df,
  param.values,
  min.sample.fraction = 0
)
}
\description{
Tests a range of models
}
