% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_param_piechart}
\alias{plot_param_piechart}
\title{Creates piechart summaries from survival summaries generated by HR_summary_for_all}
\usage{
plot_param_piechart(dinput, param, average = TRUE, add.param = FALSE)
}
\arguments{
\item{param}{the name of the parameter to summarize}

\item{average}{if TRUE, the estimates of the paramater across different runs
are averaged}

\item{add.param}{If TRUE, in case when the parameter value is not available,
0's are added.}

\item{data}{summary output created by HR_summary_for_all}
}
\description{
Creates piechart summaries from survival summaries generated by HR_summary_for_all
}
